Extract KtTypes
