+++ kotlin/reflect/src/main/java/com/squareup/moshi/kotlin/reflect/KotlinJsonAdapter.kt
@@ -437,68 +437,6 @@ class KotlinJsonAdapterFactory : JsonAdapter.Factory {
   }
 }
 
-private val KmType.isNullable: Boolean get() = Flag.Type.IS_NULLABLE(flags)
-private val KmType.canonicalName: String
-  get() {
-    return buildString {
-      val classifierString = when (val cl = classifier) {
-        is KmClassifier.Class -> createClassName(cl.name)
-        is TypeAlias -> createClassName(cl.name)
-        is TypeParameter -> arguments[cl.id].type?.canonicalName ?: "TypeVar(${cl.id})"
-      }
-      append(classifierString)
-
-      val args = arguments.joinToString(", ") {
-        "${it.variance?.name} ${it.type?.canonicalName ?: "*"}"
-      }
-
-      if (args.isNotBlank()) {
-        append('<')
-        append(args)
-        append('>')
-      }
-
-      // TODO not sure if we care about expressing the other type information here
-    }
-  }
-
-/**
- * Creates a canonical class name as represented in Metadata's [kotlinx.metadata.ClassName], where
- * package names in this name are separated by '/' and class names are separated by '.'.
- *
- * Example ClassName that we want to canonicalize: `"org/foo/bar/Baz.Nested"`.
- *
- * Local classes are prefixed with ".", but for Moshi's use case we don't deal with those.
- */
-private fun createClassName(kotlinMetadataName: String): String {
-  require(!kotlinMetadataName.isLocal) {
-    "Local/anonymous classes are not supported!"
-  }
-  return kotlinMetadataName.replace("/", ".")
-}
-
-@Suppress("SameParameterValue")
-private fun String.substringAfterLast(vararg delimiters: Char): String {
-  val index = lastIndexOfAny(delimiters)
-  return if (index == -1) this else substring(index + 1, length)
-}
-
-private fun defaultPrimitiveValue(type: Type): Any? =
-  if (type is Class<*> && type.isPrimitive) {
-    when (type) {
-      Boolean::class.java -> false
-      Char::class.java -> 0.toChar()
-      Byte::class.java -> 0.toByte()
-      Short::class.java -> 0.toShort()
-      Int::class.java -> 0
-      Float::class.java -> 0f
-      Long::class.java -> 0L
-      Double::class.java -> 0.0
-      Void.TYPE -> throw IllegalStateException("Parameter with void type is illegal")
-      else -> throw UnsupportedOperationException("Unknown primitive: $type")
-    }
-  } else null
-
 private fun Class<*>.header(): KotlinClassHeader? {
   val metadata = getAnnotation(KOTLIN_METADATA) ?: return null
   return with(metadata) {
@@ -586,106 +524,5 @@ private class JvmSignatureSearcher(clazz: Class<*>) {
     }
 
     return null
-  }
-}
-
-internal data class ParameterData(
-  val km: KmValueParameter,
-  val index: Int,
-  val rawType: Class<*>,
-  val annotations: List<Annotation>
-) {
-  val name get() = km.name
-  val declaresDefaultValue get() = Flag.ValueParameter.DECLARES_DEFAULT_VALUE(km.flags)
-  val isNullable get() = km.type!!.isNullable
-}
-
-internal data class ConstructorData(
-  val type: Class<*>,
-  val km: KmConstructor,
-  val jvm: Constructor<*>,
-  val parameters: List<ParameterData>,
-  val isDefault: Boolean
-) {
-  init {
-    jvm.isAccessible = true
-  }
-
-  fun <R> callBy(args: Map<ParameterData, Any?>): R {
-    val arguments = ArrayList<Any?>(parameters.size)
-    var mask = 0
-    val masks = ArrayList<Int>(1)
-    var index = 0
-
-    @Suppress("UseWithIndex")
-    for (parameter in parameters) {
-      if (index != 0 && index % Integer.SIZE == 0) {
-        masks.add(mask)
-        mask = 0
-      }
-
-      when {
-        args.containsKey(parameter) -> {
-          arguments.add(args[parameter])
-        }
-        parameter.declaresDefaultValue -> {
-          arguments += defaultPrimitiveValue(parameter.rawType)
-          mask = mask or (1 shl (index % Integer.SIZE))
-        }
-        else -> {
-          throw IllegalArgumentException(
-            "No argument provided for a required parameter: $parameter"
-          )
-        }
-      }
-
-      index++
-    }
-
-    if (!isDefault) {
-      @Suppress("UNCHECKED_CAST")
-      return jvm.newInstance(*arguments.toTypedArray()) as R
-    }
-
-    masks += mask
-    arguments.addAll(masks)
-
-    // DefaultConstructorMarker
-    arguments.add(null)
-
-    @Suppress("UNCHECKED_CAST")
-    return jvm.newInstance(*arguments.toTypedArray()) as R
-  }
-}
-
-internal data class PropertyData(
-  val km: KmProperty,
-  val jvmField: Field?,
-  val jvmGetter: Method?,
-  val jvmSetter: Method?,
-  val jvmAnnotationsMethod: Method?,
-  val parameter: ParameterData?
-) {
-  init {
-    jvmField?.isAccessible = true
-    jvmGetter?.isAccessible = true
-    jvmSetter?.isAccessible = true
-  }
-
-  val name get() = km.name
-
-  val javaType = jvmField?.genericType
-    ?: jvmGetter?.genericReturnType
-    ?: jvmSetter?.genericReturnType
-    ?: error("No type information available for property '${km.name}' with type '${km.returnType.canonicalName}'.")
-
-  val annotations: Set<Annotation> by lazy {
-    val set = LinkedHashSet<Annotation>()
-    jvmField?.annotations?.let { set += it }
-    jvmGetter?.annotations?.let { set += it }
-    jvmSetter?.annotations?.let { set += it }
-    jvmAnnotationsMethod?.annotations?.let { set += it }
-    parameter?.annotations?.let { set += it }
-    set
   }
 }
